pyhub-02/drf-api-01/requirements.txt
django-extensions
djangorestframework
